+++ frontend/src/Aside/CustomDatePicker.jsx
@@ -2,17 +2,18 @@ import React, {useContext, useEffect, useState} from "react";
 import DatePicker from "react-datepicker";
 import moment from 'moment';
 import "react-datepicker/dist/react-datepicker.css";
-import {HomeContext} from "./HomeContext";
+import {AppContext} from "../AppContext";
+
 const CustomDatePicker = () => {
 
-    const { movieInfo, setMovieInfo } = useContext(HomeContext);
+    const { date, setDate } = useContext(AppContext)
     const [startDate, setStartDate] = useState(moment().format("YYYY-MM-DD"));
     const onChange = (start) => {
         setStartDate(moment(start).format("YYYY-MM-DD"));
     };
 
     useEffect(() => {
-        setMovieInfo({ ...movieInfo, date: startDate });
+        setDate(startDate);
     }, [startDate]);
 
     return (
