+++ frontend/src/Aside/RegionSelector.jsx
@@ -1,7 +1,7 @@
 import {useContext, useEffect, useState} from "react";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Form from 'react-bootstrap/Form';
-import {HomeContext} from "./HomeContext";
+import {AppContext} from "../AppContext";
 
 const areas = [
     {
@@ -330,7 +330,7 @@ const areas = [
 
 const RegionSelector = () => {
 
-    const { movieInfo, setMovieInfo } = useContext(HomeContext);
+    const { region, setRegion } = useContext(AppContext)
     const [wideArea, setWideArea] = useState("");
     const [basicArea, setBasicArea] = useState("");
     const [basicAreas, setBasicAreas] = useState([]);
@@ -346,7 +346,7 @@ const RegionSelector = () => {
 
     useEffect(() => {
         setBasicAreas(areas.find((area) => area.wideArea === wideArea)?.basicArea || []);
-        setMovieInfo({ ...movieInfo, region: (wideArea + " " + basicArea).trim() });
+        setRegion((wideArea + " " + basicArea).trim());
         console.log("wideArea : ", wideArea);
         console.log("basicArea : ", basicArea);
     }, [wideArea, basicArea]);
